第一关大体框架
完成生成随机阳光
指定僵尸的出场时间
收集阳光的监听器
显示阳光值

+++ Plants_VS_Zombies/Classes/Zombie.cpp
@@ -27,12 +27,12 @@ bool Zombie::getHeadCondition()
 	return lostHead;
 }
 
-void Zombie::setStartTime(int time)
+void Zombie::setStartTime(float time)
 {
 	startTime = time;
 }
 
-int Zombie::getStartTime()
+float Zombie::getStartTime()
 {
 	return startTime;
 }
